+++ src/main/java/org/example/Auto.java
@@ -1,0 +1,12 @@
+package org.example;
+
+public class Auto {
+    String brand;
+    int speed;
+
+    Auto(String brand, int speed) {
+
+        this.brand = brand;
+        this.speed = speed;
+    }
+}
